# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/change-status-items": {
            "post": {
                "description": "change status items by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Change status items",
                "parameters": [
                    {
                        "description": "Change status by listID",
                        "name": "ID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ListIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cập nhật dữ liệu thành công",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Binding dữ liệu lỗi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Dữ liệu không tồn tại",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Tìm kiếm dữ liệu lỗi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/delete-items": {
            "post": {
                "description": "Delete items by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Delete Items",
                "parameters": [
                    {
                        "description": "Delete items by listID",
                        "name": "ID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ListIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa dữ liệu thành công",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Binding dữ liệu lỗi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Dữ liệu không tồn tại",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Tìm kiếm dữ liệu lỗi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/export-items": {
            "get": {
                "description": "export excel list items form the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "export list items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ngày bắt đầu lấy dữ liệu theo format YYYY-MM-DD",
                        "name": "from-date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ngày kết thúc lấy dữ liệu theo format YYYY-MM-DD",
                        "name": "to-date",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Trạng thái item, true hoặc false",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Từ khóa tìm kiếm theo tiêu đề item",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trả về file excel",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Trạng thái item tìm kiếm không hợp lệ",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Lấy dữ liệu hoặc tạo file excel lỗi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/export-pdf": {
            "post": {
                "description": "Converts html to a PDF file using wkhtmltopdf library and returns the PDF file.",
                "consumes": [
                    "application/html"
                ],
                "produces": [
                    "application/pdf"
                ],
                "summary": "Export an HTML file to a PDF file.",
                "parameters": [
                    {
                        "description": "HTML file to be converted to PDF",
                        "name": "html",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file as an attachment",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item": {
            "post": {
                "description": "Create a new Item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create a Item",
                "parameters": [
                    {
                        "description": "New Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collections.Item"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tạo dữ liệu thành công",
                        "schema": {
                            "$ref": "#/definitions/collections.Item"
                        }
                    },
                    "400": {
                        "description": "Dữ liệu gửi lên không chính xác",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Tiêu đề không được bỏ trống",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Tạo item lỗi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item/{id}": {
            "put": {
                "description": "Update an existing item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Update an item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item object that needs to update",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collections.Item"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cập nhật dữ liệu thành công",
                        "schema": {
                            "$ref": "#/definitions/collections.Item"
                        }
                    },
                    "400": {
                        "description": "Binding lỗi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Dữ liệu không tồn tại",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Tiêu đề không được bỏ trống",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Cập nhật dữ liệu lỗi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/items": {
            "get": {
                "description": "get list items form the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "get list items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ngày bắt đầu lấy dữ liệu theo format YYYY-MM-DD",
                        "name": "from-date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ngày kết thúc lấy dữ liệu theo format YYYY-MM-DD",
                        "name": "to-date",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Trạng thái item, true hoặc false",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Từ khóa tìm kiếm theo tiêu đề item",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Trạng thái item tìm kiếm không hợp lệ",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy dữ liệu",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Tìm kiếm dữ liệu lỗi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "collections.Item": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "controllers.ListIDRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}